CloseableChannel split to search and fill inokers
